# REST Server/target/docs/apidocs/ui/swagger.json
{
  "swagger": "2.0",
  "info" : {
    "description" : "<h1>MavenServer<\/h1>",
    "version" : "1.3",
    "title" : "MavenServer"
  },
  "schemes" : [],
...
  "paths": {
...
    "\/search/subfolder123/{id}" : {
      "post" : {
        "tags" : [ "Search" ],
        "description" : "",
        "operationId" : "resource_Search_postTest_POST",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [
          {
            "name" : "id",
            "in" : "path",
            "required" : true,
            "type" : "integer",
            "description" : ""
          },
          {
            "name" : "body",
            "in" : "body",
            "schema" : {
"description" : "",
"type" : "object",
"additionalProperties" : {
  "type" : "file"
}
            },
            "description" : ""
          }
        ],
        "responses" : {
          "201" : {
            "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "type" : "file"
}
            },
            "headers" : {
            },
            "examples" : {
            },
            "description" : "Success"
          }
        }
      }
    }
...
    "\/search/zip/{zip}" : {
      "get" : {
        "tags" : [ "Search" ],
        "description" : "",
        "operationId" : "resource_Search_getZip_GET",
        "produces" : [ "application/xml" ],
        "parameters" : [
          {
            "name" : "zip",
            "in" : "path",
            "required" : true,
            "type" : "string",
            "description" : ""
          }
        ],
        "responses" : {
          "200" : {
            "schema" : {
"description" : "",
"type" : "string"
            },
            "headers" : {
            },
            "examples" : {
            },
            "description" : "Success"
          }
        }
      }
    }
...
  }
}